chatbot_app/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-jose
passlib
python-dotenv
langchain-groq
langgraph
psycopg2-binary
